backend/django_hstat/requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==5.0.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-sslserver==0.22
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
mysqlclient==2.2.4
packaging==24.0
PyJWT==2.8.0
pytz==2024.1
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
whitenoise==6.6.0
